# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5f512d2467541c707d17b59d
  title: TESTAPP
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /registrationinfo:
    post:
      x-skaffolder-id: 5f5139e967541c707d17b681
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5f51305b67541c707d17b649
      x-skaffolder-resource: UserInfo
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - UserInfo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Address:
                  type: string
                Company:
                  type: boolean
                Country:
                  type: string
                Firstname:
                  type: string
                Lastname:
                  type: string
                WebsiteAddr:
                  type: string
    get:
      x-skaffolder-id: 5f5139e967541c707d17b687
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5f51305b67541c707d17b649
      x-skaffolder-resource: UserInfo
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - UserInfo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserInfo"
  "/registrationinfo/{id}":
    delete:
      x-skaffolder-id: 5f5139e967541c707d17b688
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5f51305b67541c707d17b649
      x-skaffolder-resource: UserInfo
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - UserInfo
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5f5139e967541c707d17b684
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5f51305b67541c707d17b649
      x-skaffolder-resource: UserInfo
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - UserInfo
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
    post:
      x-skaffolder-id: 5f5139e967541c707d17b682
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5f51305b67541c707d17b649
      x-skaffolder-resource: UserInfo
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - UserInfo
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInfo"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserInfo:
      x-skaffolder-id: 5f51305b67541c707d17b649
      x-skaffolder-id-db: 5f512d2467541c707d17b59e
      x-skaffolder-id-entity: 5f51305b67541c707d17b641
      x-skaffolder-url: /registrationinfo
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Address:
          type: string
          x-skaffolder-id-attr: 5f51305b67541c707d17b644
          x-skaffolder-type: String
          x-skaffolder-required: true
        Company:
          type: boolean
          x-skaffolder-id-attr: 5f51305b67541c707d17b647
          x-skaffolder-type: Boolean
        Country:
          type: string
          x-skaffolder-id-attr: 5f51305b67541c707d17b645
          x-skaffolder-type: String
          x-skaffolder-required: true
        Firstname:
          type: string
          x-skaffolder-id-attr: 5f51305b67541c707d17b643
          x-skaffolder-type: String
        Lastname:
          type: string
          x-skaffolder-id-attr: 5f51305b67541c707d17b642
          x-skaffolder-type: String
          x-skaffolder-required: true
        WebsiteAddr:
          type: string
          x-skaffolder-id-attr: 5f51305b67541c707d17b648
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - Address
        - Country
        - Lastname
        - WebsiteAddr
...